WindowBar
- Create a new WindowBar folder.
- Gather all the window bar scripts
- Add an assembly def for window bar targetting Windows
