--- UMI3D-BROWSER/Assets/PC/Project/Runtime/PC.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2ae292775153e1f45900f8da913323a8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
