+++ UMI3D-BROWSER/Assets/Runtime/WindowBar/01-Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9257980b2615dac418ddfb16755fa1bf
+guid: 2a15a979841687042b66836a984048df
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
