+++ UMI3D-BROWSER/Assets/Runtime/WindowBar/01-Scripts/HideApplicationButton.cs
