+++ UMI3D-BROWSER/Assets/Runtime/WindowBar/01-Scripts/WindowBarPanel.cs
@@ -14,24 +14,28 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-using umi3d.browserRuntime.pc;
+using inetum.unityUtils;
+using umi3d.browserRuntime.notificationKeys;
 using UnityEngine;
 
 namespace umi3d.browserRuntime.ui.windowBar
 {
-    public class WindowBarUpdate : MonoBehaviour
+    public class WindowBarPanel : MonoBehaviour
     {
         private void Awake()
         {
 #if UNITY_EDITOR
-            gameObject.SetActive(false);
+            transform.parent.gameObject.SetActive(false);
             return;
 #endif
+            NotificationHub.Default.Subscribe(this, WindowsManagerNotificationKey.FullScreenModeChanged, OnFullScreenModeChanged);
+            transform.parent.gameObject.SetActive(WindowsManager.IsWindowInFullScreen);
         }
 
-        private void Update()
+        private void OnFullScreenModeChanged(Notification notification)
         {
-            WindowsManager.Update();
+            if (notification.TryGetInfoT<FullScreenMode>(WindowsManagerNotificationKey.FullScreenModeChangedInfo.Mode, out var mode))
+                transform.parent.gameObject.SetActive(mode != FullScreenMode.Windowed);
         }
     }
 }
