+++ UMI3D-BROWSER/Assets/Runtime/WindowBar/01-Scripts/WindowBarUpdate.cs
@@ -14,30 +14,23 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-using inetum.unityUtils;
-using System;
-using umi3d.browserRuntime.notificationKeys;
-using umi3d.browserRuntime.pc;
 using UnityEngine;
 
 namespace umi3d.browserRuntime.ui.windowBar
 {
-    public class WindowBarPanel : MonoBehaviour
+    public class WindowBarUpdate : MonoBehaviour
     {
         private void Awake()
         {
 #if UNITY_EDITOR
-            transform.parent.gameObject.SetActive(false);
+            gameObject.SetActive(false);
             return;
 #endif
-            NotificationHub.Default.Subscribe(this, WindowsManagerNotificationKey.FullScreenModeChanged, OnFullScreenModeChanged);
-            transform.parent.gameObject.SetActive(WindowsManager.IsWindowInFullScreen);
         }
 
-        private void OnFullScreenModeChanged(Notification notification)
+        private void Update()
         {
-            if (notification.TryGetInfoT<FullScreenMode>(WindowsManagerNotificationKey.FullScreenModeChangedInfo.Mode, out var mode))
-                transform.parent.gameObject.SetActive(mode != FullScreenMode.Windowed);
+            WindowsManager.Update();
         }
     }
 }
