+++ UMI3D-BROWSER/Assets/Runtime/WindowBar/umi3d.browserRuntime.ui.windowBar.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "umi3d.browserRuntime.pc",
+    "name": "umi3d.browserRuntime.ui.windowBar",
     "rootNamespace": "",
     "references": [
         "GUID:d0bf4ea2ef3a7c74e9160b8995fd902b",
